test(core): extract RISC-V system creation to separate header
Follow same pattern as Pep/10, where one helper is used across many tests.
Starting to work on load/store instruction tests, which also need a CPU.
